smcp-gmail-imap/requirements.txt
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
requests>=2.28.0
